Safari-609.1.11.1/LayoutTests/fast/canvas/webgl/webgl-debug-renderer-info-expected.txt

CONSOLE MESSAGE: line 97: WebGL: INVALID_ENUM: getParameter: invalid parameter name, WEBGL_debug_renderer_info not enabled
CONSOLE MESSAGE: line 101: WebGL: INVALID_ENUM: getParameter: invalid parameter name, WEBGL_debug_renderer_info not enabled
This test verifies the functionality of the WEBGL_debug_renderer_info extension, if it is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists
Testing enums with extension disabled
PASS getError was expected value: INVALID_ENUM : UNMASKED_VENDOR_WEBGL should not be queryable if extension is disabled
PASS getError was expected value: INVALID_ENUM : UNMASKED_RENDERER_WEBGL should not be queryable if extension is disabled
PASS Successfully enabled WEBGL_debug_renderer_info extension
PASS WEBGL_debug_renderer_info listed as supported and getExtension succeeded
Testing enums with extension enabled
PASS ext.UNMASKED_VENDOR_WEBGL is 0x9245
PASS getError was expected value: NO_ERROR : UNMASKED_VENDOR_WEBGL query should succeed if extension is enable
PASS ext.UNMASKED_RENDERER_WEBGL is 0x9246
PASS getError was expected value: NO_ERROR : UNMASKED_RENDERER_WEBGL query should succeed if extension is enable

PASS successfullyParsed is true

TEST COMPLETE
